Move IDEEditor and IDEMenu into the components folder
