+++ game_frontend/src/components/GamePage/index.js
@@ -4,13 +4,13 @@ import IDE from 'containers/IDE'
 import Game from 'components/Game'
 
 export const GamePageLayout = styled.div`
-  display: grid
-  grid-template: 80px 1fr 250px / 1fr 1fr
+  display: grid;
+  grid-template: 80px 1fr 250px / 1fr 1fr;
   grid-template-areas: "ide-menu game-menu"
                        "ide-editor game-view"
                        "ide-console game-view";
-  width: 100vw
-  height: 100vh
+  width: 100vw;
+  height: 100vh;
 `
 
 export default class GamePage extends Component {
